+++ tests/store/test_tag_filters.py
@@ -1,6 +1,6 @@
 from housekeeper.store import Store
 from housekeeper.store.models import Tag
-from housekeeper.store.tag_filters import (
+from housekeeper.store.filters.tag_filters import (
     filter_tag_by_name,
     TagFilter,
     apply_tag_filter,
